% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{methods}
\name{concordance}
\alias{concordance}
\alias{concordance,list-method}
\title{Calculate the concordance between the vectorial elements of a list.}
\usage{
concordance(obj, ...)

\S4method{concordance}{list}(obj, ...)
}
\arguments{
\item{obj}{a list of vectors of the same type}

\item{...}{additional parameters}
}
\description{
Calculate the concordance between the vectorial elements of a list.
}

