% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{drop}
\alias{drop}
\title{Drops columns that are not needed}
\usage{
drop(obj, ...)
}
\arguments{
\item{obj}{object that should be dropped}

\item{...}{additional params}
}
\description{
Drops columns that are not needed
}

